app/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-easy-validator==1.7.0
django-cors-headers==4.3.1
environs==11.0.0
PyJWT==2.8.0
requests==2.31.0
